improved

+++ zips/install.sh
@@ -2,11 +2,15 @@
 
 version=005
 
-unzip mlclass-ex\*
-mv mlclass-ex*/* .
-rm -rf mlclass-ex?-$version
-mv ex*.pdf ..
-mv mlclass-ex? ..
-rm -rf __MACOSX/
+md5sum --check --status checksums-$version
+if [ $? -eq 0 ]; then
+  unzip mlclass-ex\*
+  mv mlclass-ex*/* .
+  rm -rf mlclass-ex?-$version
+  mv ex*.pdf ..
+  mv mlclass-ex? ..
+  rm -rf __MACOSX/
+else
+  echo 'checksum fail, aborting'
+fi
 
-echo $version
